fix: resolve minor bugs

+++ silogistik2106751732/src/main/java/apap/ti/silogistik2106751732/DTO/response/HomeResponseDTO.java
@@ -8,8 +8,8 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Data
 public class HomeResponseDTO {
-    private Integer jumlahGudang;
-    private Integer jumlahBarang;
-    private Integer jumlahPermintaanPengiriman;
-    private Integer jumlahKaryawan;
+    private Long jumlahGudang;
+    private Long jumlahBarang;
+    private Long jumlahPermintaanPengiriman;
+    private Long jumlahKaryawan;
 }
